llvm: Move module compilation to jit_engine
